Moved most arguments to seperate configuration file.
In order to run data from multiple counties, with different parameters,
in the same simulation, the software needs to be aware of what agent belongs
to what county.

This means the input files need to be grouped by county (which also speeds
up the reading of the files due to the cross-referencing betwen files), for
which a configuration file is the easiest option.