Bug fixes and efficiency improvements.
Fixed huge bug that agents would apply a norm only when they decided to ignore it (oops).
Fixed bug introduced by adding the logging files which could crash the agent code due to
concurrency issues.
Fixed bug in structure of logging files.

Removed end-of-simulation logging of all agent's radii of gyration, as this was very
memory hungry, and no longer necessary now we keep track of what county an agent belongs
to