Hopefully improved multithreading
Changed the phases of processing (at the start of a time step) and creating (at
the end) of both data frames multithreaded, by naively assigning a batch to
each available thread. Except for the visits data frames, processing a single
row should be the same time complexity for each row. For the visits data frames,
this is not necessarily the case, as one agent may have multiple actions. Hopefully
this balances out with a large enough number of agents.

Writing to an arrow data frame cannot occur parallel with random access (probably
due to the buffer used), but writing to distinct "chuncks" of the data frame does work,
which means that for now the naive batch approach is best.

A few minor changes to Sim2APL have been made, to do less processing on the agent actions.
This means changes to how those actions are processed in this simulation have been updated
as well.

Two temporary classes have been added to time the execution of the various phases within
a deliberation step. Perhaps later these will be reverted to the Sim2APL default implementations.

I now close the data frames after they have been processed. Hopefully this will have a positive
impact on memory usage.