Made all agents share the same norm context.
Previously, all agents had their own norm context, which resulted
in n classes for n agents, with n updates for each changed norm in
a time step.
Now, one normcontext object is passed-by-reference to all agents,
and norm updates are O(1).

Made changes to Sim2APL to how agent actions are returned final and
updated the code in this library to reflect those changes.
PostTickHook processing is still not being performed, as this requires
additional optimization